+++ inform/Project/Toolbar/IFToolbarStatusView.h
@@ -21,8 +21,8 @@
 
 -(void) canCancel: (BOOL) canCancel;
 -(void) startProgress;
--(void) setProgressMaxValue: (float) maxValue;
--(void) updateProgress: (float) progress;
+-(void) setProgressMaxValue: (CGFloat) maxValue;
+-(void) updateProgress: (CGFloat) progress;
 -(void) setProgressIndeterminate: (BOOL) indeterminate;
 -(void) stopProgress;
